[NV-8287] global ignore
