+++ tests/test_spell_views.py
@@ -108,24 +108,3 @@ def test_with_exception_word_in_dictionary(
     # and than check that excepted word not in the check output
     server_response = run_request()
     assert tested_word not in parse_words(server_response)
-
-
-@pytest.mark.parametrize(
-    ("wannabe_user_input", "excluded_words"),
-    [("ШЯЧЛО ПОПЯЧТСА ПОПЯЧТСА", "шЯчЛо, ПоПяЧтСа")],
-)
-def test_default_excluded_words(
-    app_client: "TestClient",
-    wannabe_user_input: str,
-    excluded_words: str,
-    monkeypatch: typing.Any,
-) -> None:
-    """Dead simple test."""
-    with monkeypatch.context() as patcher:
-        patcher.setattr(SETTINGS, "exclusion_words", excluded_words)
-        server_response: typing.Final = app_client.post(
-            f"{SETTINGS.api_prefix}/check/",
-            json=models.SpellCheckRequest(text=wannabe_user_input, language=RU_LANG).model_dump(),
-        )
-        assert server_response.status_code == 200
-        assert server_response.json()["corrections"] == [], f"{server_response.json()=} --- {excluded_words=}"
