+++ whole_app/spell.py
@@ -31,13 +31,13 @@ class SpellCheckService:
         """Initialize machinery."""
         self._input_text = request_payload.text
         self._exclusion_words = exclusion_words if exclusion_words else []
+        self._exclusion_words.extend(SETTINGS.exclusion_words_set)
 
         if request_payload.exclude_urls:
             for one_url in self._url_extractor.find_urls(self._input_text):
                 self._exclusion_words.extend(
-                    [word.lower() for word in re.split(SEPARATORS_TO_SPLIT_URL_BY_WORDS, one_url)]
+                    {one_word.lower() for one_word in re.split(SEPARATORS_TO_SPLIT_URL_BY_WORDS, one_url)}
                 )
-
         self._spellcheck_engine = SpellChecker(request_payload.language)
         return self
