+++ whole_app/views.py
@@ -63,7 +63,6 @@ async def spell_check_main_endpoint(
         exclusion_words = await storage_engine.prepare(
             request_payload.user_name,
         ).fetch_records()
-    exclusion_words.extend([one_word.strip().lower() for one_word in SETTINGS.exclusion_words.split(",")])
     return models.SpellCheckResponse(
         **request_payload.model_dump(),
         corrections=await to_thread.run_sync(
